refact(CustomerPage): break down components
